[cbackend] Rename CJuvix to MiniC
